logictest: improve handling of results with whitespace
Previously, rowsort would fail to sort queries that returned results
with whitespace. For example, the following logic test would reliably
fail:

    query TT rowsort
    VALUES ('Multi line', 'strings'), ('Used to break', 'rowsort') ORDER BY 1
    ----
    Used to break  rowsort
    Multi line     strings

This commit teaches rowsort to handle such results.

Note that whitespace in logictests has odd semantics: a row of two
strings, like ('Used to break', 'rowsort'), is internally parsed into a
row of five strings, like ('Used', 'to', 'break', 'rowsort'). This means
that it is not possible to support partialsort when a result contains
whitespace, as partialsort needs to be able to unambiguously determine
column boundaries, so this additionally adds an error message if an
improper use of partial sort is detected.

Release note: None